# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mediasearch",
    "summary": "EmbeddingGemma 2 を使った録画映像・音声・画像のマルチモーダル検索 API(開発用)",
    "description": "\n録画映像・音声・画像を取り込み、**文章・画像・音声**で該当する場面を探すための API です。\n\n- 取り込み(`/api/ingest/*`)はバックグラウンドのジョブで行い、応答の `job_id` で進み具合を確認します\n  (`GET /api/jobs/{job_id}`)。\n- 検索(`/api/search/*`)は、スコア(コサイン類似度)の高い順に窓(時間の区切り、または静止画 1 枚)を返します。\n- 元ファイルは `GET /api/media/{source_id}`(Range 対応)、サムネイルは `GET /api/thumb/{window_id}` で取得します。\n- エラーは `{\"detail\": \"...\"}` の形で返します(説明は日本語)。\n\n**認証**: サーバに `API_TOKEN` が設定されている場合、`Authorization: Bearer <トークン>` か\n`X-API-Key: <トークン>` が必要です\n(右上の Authorize から設定できます)。設定されていない場合は不要です。\n\nAI エージェントからは、同じ機能を MCP(`/mcp`)で利用できます。\n開発者向けの説明は docs/api.md と docs/mcp.md を参照してください。\n",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/compute/info": {
      "get": {
        "tags": [
          "compute"
        ],
        "summary": "推論側の情報",
        "operationId": "info_compute_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeInfo"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/compute/embed/texts": {
      "post": {
        "tags": [
          "compute"
        ],
        "summary": "文章をベクトルにする",
        "operationId": "embed_texts_compute_embed_texts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vectors"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/compute/embed/images": {
      "post": {
        "tags": [
          "compute"
        ],
        "summary": "画像をベクトルにする",
        "operationId": "embed_images_compute_embed_images_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "files": {
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "type": "array",
                    "description": "画像(JPEG など)"
                  },
                  "dims": {
                    "type": "integer"
                  },
                  "high": {
                    "type": "string",
                    "description": "1 なら検索用の優先度で処理"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vectors"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる(MAX_UPLOAD_MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/compute/embed/audio": {
      "post": {
        "tags": [
          "compute"
        ],
        "summary": "音声をベクトルにする",
        "operationId": "embed_audio_compute_embed_audio_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "16bit PCM・16kHz・モノラル(リトルエンディアン)"
                  },
                  "dims": {
                    "type": "integer"
                  },
                  "high": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vectors"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる(MAX_UPLOAD_MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/compute/embed/parts": {
      "post": {
        "tags": [
          "compute"
        ],
        "summary": "文字・画像・音声を並べた順序のまま 1 ベクトルにする",
        "description": "文字・画像・音声を、並べた順序のまま1ベクトルにします。",
        "operationId": "embed_parts_compute_embed_parts_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "manifest": {
                    "type": "string",
                    "description": "JSON。[{\"type\": \"text\", \"text\": \"00:00\"}, {\"type\": \"image\", \"file\": \"f0\"}, ...]"
                  },
                  "dims": {
                    "type": "integer"
                  },
                  "high": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vectors"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "リクエストが大きすぎる(MAX_UPLOAD_MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/info": {
      "get": {
        "tags": [
          "状態"
        ],
        "summary": "状態・版・既定値・索引の件数",
        "description": "推論側(モデル・デバイス)、アプリの版、取り込みの既定値とプリセット、索引の件数、監視フォルダの状態を返します。",
        "operationId": "info_api_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
...
    "/api/search/text": {
      "post": {
        "tags": [
          "検索"
        ],
        "summary": "文章で探す",
        "description": "文章に近い場面(窓)を、スコアの高い順に返します。`kind=auto` は映像の窓と静止画を探します。",
        "operationId": "search_text_api_search_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "入力が正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DB のモデル・次元と、現在の設定が一致しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "推論側(compute)に接続できない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComputeInfo": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "dims": {
            "type": "integer",
            "title": "Dims"
          },
          "dtype": {
            "type": "string",
            "title": "Dtype"
          },
          "device": {
            "type": "string",
            "title": "Device"
          },
          "accelerator": {
            "type": "string",
            "title": "Accelerator"
          },
          "modalities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modalities"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "model_id",
          "dims",
          "dtype",
          "device",
          "accelerator",
          "modalities",
          "backend"
        ],
        "title": "ComputeInfo"
      },
...
      "EmbedderStatus": {
        "properties": {
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend",
            "description": "local / remote / dummy"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "dims": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dims",
            "description": "保存するベクトルの次元"
          },
          "dtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtype"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "accelerator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accelerator",
            "description": "cpu / nvidia-cuda / amd-rocm / intel-xpu / dummy"
          },
          "modalities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modalities"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "推論側に接続できない場合の理由"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "EmbedderStatus"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "description": "エラーの内容(日本語の説明。入力の検証エラーでは項目ごとの一覧)"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "examples": [
          {
            "detail": "取り込み元が見つかりません"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexStatus": {
        "properties": {
          "sources": {
            "type": "integer",
            "title": "Sources"
          },
          "windows": {
            "type": "integer",
            "title": "Windows"
          },
          "windows_by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Windows By Kind"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id",
            "description": "DB を作ったときのモデル(未作成なら null)"
          },
          "dims": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dims"
          },
          "queue_size": {
            "type": "integer",
            "title": "Queue Size",
            "description": "取り込みを待っているジョブの数"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "sources",
          "windows",
          "windows_by_kind",
          "queue_size"
        ],
        "title": "IndexStatus"
      },
      "InfoResponse": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "version": {
            "$ref": "#/components/schemas/VersionInfo"
          },
          "embedder": {
            "$ref": "#/components/schemas/EmbedderStatus"
          },
          "defaults": {
            "additionalProperties": true,
            "type": "object",
            "title": "Defaults",
            "description": "取り込みの既定値(窓の長さなど)"
          },
          "presets": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Presets",
            "description": "取り込みのプリセット"
          },
          "index": {
            "$ref": "#/components/schemas/IndexStatus"
          },
          "watch": {
            "additionalProperties": true,
            "type": "object",
            "title": "Watch",
            "description": "監視フォルダの状態"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "role",
          "version",
          "embedder",
          "defaults",
          "presets",
          "index",
          "watch"
        ],
        "title": "InfoResponse"
      },
...
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Results",
            "description": "スコアの高い順の窓"
          },
          "intervals": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SearchInterval"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervals",
            "description": "同じ取り込み元で隣り合う窓をまとめた区間(merge=true のとき)"
          },
          "took_ms": {
            "type": "integer",
            "title": "Took Ms",
            "description": "DB の検索にかかった時間(ミリ秒)"
          },
          "embed_ms": {
            "type": "integer",
            "title": "Embed Ms",
            "description": "クエリのベクトル化にかかった時間(ミリ秒)"
          },
          "searched_kinds": {
            "type": "string",
            "title": "Searched Kinds",
            "description": "指定された kind"
          }
        },
        "type": "object",
        "required": [
          "results",
          "took_ms",
          "embed_ms",
          "searched_kinds"
        ],
        "title": "SearchResponse"
      },
...
      "TextSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query",
            "description": "探したい内容を表す文章(日本語・英語など)"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "description": "返す件数(省略時は TOP_K)"
          },
          "min_score": {
            "type": "number",
            "title": "Min Score",
            "description": "これより低いスコアは返さない(該当なしを判定するため)",
            "default": 0.0
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id",
            "description": "グループ ID で絞り込む"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "場所で絞り込む"
          },
          "from_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Ts",
            "description": "この日時以降(UNIX 秒か ISO 8601)"
          },
          "to_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Ts",
            "description": "この日時以前(UNIX 秒か ISO 8601)"
          },
          "kind": {
            "type": "string",
            "enum": [
              "auto",
              "all",
              "frames",
              "tav",
              "audio",
              "image"
            ],
            "title": "Kind",
            "description": "探す窓の種類",
            "default": "auto"
          },
          "merge": {
            "type": "boolean",
            "title": "Merge",
            "description": "隣り合う窓を区間にまとめた intervals も返す",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "TextSearchRequest",
        "examples": [
          {
            "from_ts": "2026-01-01T00:00:00",
            "min_score": 0.3,
            "query": "赤い車が通り過ぎる",
            "top_k": 10
          }
        ]
      },
      "TextsRequest": {
        "properties": {
          "texts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Texts",
            "description": "ベクトル化する文章"
          },
          "kind": {
            "type": "string",
            "enum": [
              "query",
              "document"
            ],
            "title": "Kind",
            "description": "検索文(query)か、検索される側の文書(document)か",
            "default": "query"
          },
          "dims": {
            "type": "integer",
            "title": "Dims",
            "description": "返す次元(768 / 512 / 256 / 128)",
            "default": 768
          }
        },
        "type": "object",
        "required": [
          "texts"
        ],
        "title": "TextsRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Vectors": {
        "properties": {
          "vectors": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Vectors",
            "description": "L2 正規化済みのベクトル"
          }
        },
        "type": "object",
        "required": [
          "vectors"
        ],
        "title": "Vectors"
      },
      "VersionInfo": {
        "properties": {
          "app": {
            "type": "string",
            "title": "App",
            "description": "アプリの版"
          },
          "variant": {
            "type": "string",
            "title": "Variant",
            "description": "イメージの種類(slim / cpu / cuda / rocm / intel。ソースから動かすと source)"
          },
          "revision": {
            "type": "string",
            "title": "Revision",
            "description": "ビルド元のコミット"
          },
          "build_ref": {
            "type": "string",
            "title": "Build Ref",
            "description": "ビルド元のタグ・ブランチ"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "app",
          "variant",
          "revision",
          "build_ref"
        ],
        "title": "VersionInfo"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_TOKEN の値"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API_TOKEN の値"
      }
    }
  },
...
}